refactor(agent): split runtime state
