+++ packages/agent/src/model.ts
@@ -1,39 +1,3 @@
-export const positiveInt = (value: number): number => {
-	if (!Number.isInteger(value) || value < 1)
-		throw new Error("expected positive integer");
-	return value;
-};
-
-export const tickId = (value: number): number => {
-	if (!Number.isInteger(value) || value < 0)
-		throw new Error("expected non-negative integer");
-	return value;
-};
-
-const identifier = (value: string, name: string): string => {
-	if (
-		typeof value !== "string" ||
-		value.length === 0 ||
-		!/^[A-Za-z0-9._:-]+$/.test(value)
-	) {
-		throw new Error(`invalid ${name}`);
-	}
-	return value;
-};
-
-const nonEmpty = (value: string, name: string): string => {
-	if (typeof value !== "string" || value.length === 0)
-		throw new Error(`invalid ${name}`);
-	return value;
-};
-
-export const sourceId = (value: string): string =>
-	identifier(value, "SourceId");
-export const subjectKey = (value: string): string =>
-	nonEmpty(value, "SubjectKey");
-export const workKey = (value: string): string => nonEmpty(value, "WorkKey");
-export const runId = (value: string): string => identifier(value, "RunId");
-
 export type AgentPhase =
 	| "setup"
 	| "observe"
@@ -44,12 +8,12 @@ export type AgentPhase =
 export type HookPhase = "observe" | "reconcile" | "select";
 
 export class PlotAgentError extends Error {
-	readonly phase: AgentPhase;
-	readonly source_id?: string;
+	phase: AgentPhase;
+	source_id?: string;
 	constructor(input: {
-		readonly phase: AgentPhase;
-		readonly message: string;
-		readonly source_id?: string;
+		phase: AgentPhase;
+		message: string;
+		source_id?: string;
 	}) {
 		super(input.message);
 		this.name = "PlotAgentError";
@@ -59,14 +23,14 @@ export class PlotAgentError extends Error {
 }
 
 export interface Observation {
-	readonly type: string;
-	readonly subject?: string;
-	readonly data?: unknown;
+	type: string;
+	subject?: string;
+	data?: unknown;
 }
 export interface SetFactProposal {
-	readonly type: "set_fact";
-	readonly key: string;
-	readonly value: unknown;
+	type: "set_fact";
+	key: string;
+	value: unknown;
 }
 export const setFact = (key: string, value: unknown): SetFactProposal => ({
 	type: "set_fact",
@@ -74,37 +38,40 @@ export const setFact = (key: string, value: unknown): SetFactProposal => ({
 	value,
 });
 export interface RemoveFactProposal {
-	readonly type: "remove_fact";
-	readonly key: string;
+	type: "remove_fact";
+	key: string;
 }
 export const removeFact = (key: string): RemoveFactProposal => ({
 	type: "remove_fact",
 	key,
 });
 export interface InterruptWorkProposal {
-	readonly type: "interrupt_work";
-	readonly workKey: string;
-	readonly reason?: string;
+	type: "interrupt_work";
+	workKey: string;
+	reason?: string;
 }
 export const interruptWork = (
 	key: string,
 	reason?: string,
-): InterruptWorkProposal => ({
-	type: "interrupt_work",
-	workKey: key,
-	...(reason === undefined ? {} : { reason }),
-});
+): InterruptWorkProposal => {
+	const proposal: InterruptWorkProposal = {
+		type: "interrupt_work",
+		workKey: key,
+	};
+	if (reason !== undefined) proposal.reason = reason;
+	return proposal;
+};
 export interface ScheduleWakeOptions {
-	readonly reason?: string;
-	readonly workKey?: string;
-	readonly attempt?: number;
+	reason?: string;
+	workKey?: string;
+	attempt?: number;
 }
 export interface ScheduleWakeProposal {
-	readonly type: "schedule_wake";
-	readonly delayMs: number;
-	readonly reason?: string;
-	readonly workKey?: string;
-	readonly attempt?: number;
+	type: "schedule_wake";
+	delayMs: number;
+	reason?: string;
+	workKey?: string;
+	attempt?: number;
 }
 export const scheduleWake = (
 	delayMs: number,
@@ -114,37 +81,36 @@ export const scheduleWake = (
 		typeof reasonOrOptions === "string"
 			? { reason: reasonOrOptions }
 			: (reasonOrOptions ?? {});
-	return {
+	const proposal: ScheduleWakeProposal = {
 		type: "schedule_wake",
-		delayMs: positiveInt(delayMs),
-		...(options.reason === undefined ? {} : { reason: options.reason }),
-		...(options.workKey === undefined ? {} : { workKey: options.workKey }),
-		...(options.attempt === undefined
-			? {}
-			: { attempt: positiveInt(options.attempt) }),
+		delayMs,
 	};
+	if (options.reason !== undefined) proposal.reason = options.reason;
+	if (options.workKey !== undefined) proposal.workKey = options.workKey;
+	if (options.attempt !== undefined) proposal.attempt = options.attempt;
+	return proposal;
 };
 
 export interface WorkDisplay {
-	readonly kind?: string;
-	readonly primary?: string;
-	readonly title?: string;
-	readonly subtitle?: string;
-	readonly url?: string;
-	readonly version?: string;
-	readonly labels?: readonly string[];
+	kind?: string;
+	primary?: string;
+	title?: string;
+	subtitle?: string;
+	url?: string;
+	version?: string;
+	labels?: string[];
 }
 export interface OperatorActionConfirm {
-	readonly title: string;
-	readonly message?: string;
+	title: string;
+	message?: string;
 }
 export interface OperatorAction {
-	readonly id: string;
-	readonly label: string;
-	readonly tone?: "primary" | "secondary" | "danger";
-	readonly disabledReason?: string;
-	readonly requiresComment?: boolean;
-	readonly confirm?: OperatorActionConfirm;
+	id: string;
+	label: string;
+	tone?: "primary" | "secondary" | "danger";
+	disabledReason?: string;
+	requiresComment?: boolean;
+	confirm?: OperatorActionConfirm;
 }
 export type WorkStatus =
 	| "pending"
@@ -156,19 +122,19 @@ export type WorkStatus =
 	| "failed";
 
 export interface WorkRecord {
-	readonly workKey: string;
-	readonly sourceId: string;
-	readonly status: WorkStatus;
-	readonly subject?: string;
-	readonly display?: WorkDisplay;
-	readonly blockedReason?: string;
-	readonly operatorActions?: readonly OperatorAction[];
-	readonly currentRunId?: string;
+	workKey: string;
+	sourceId: string;
+	status: WorkStatus;
+	subject?: string;
+	display?: WorkDisplay;
+	blockedReason?: string;
+	operatorActions?: OperatorAction[];
+	currentRunId?: string;
 }
 
 export interface UpsertWorkProposal {
-	readonly type: "upsert_work";
-	readonly work: WorkRecord;
+	type: "upsert_work";
+	work: WorkRecord;
 }
 export const upsertWork = (work: WorkRecord): UpsertWorkProposal => ({
 	type: "upsert_work",
@@ -176,8 +142,8 @@ export const upsertWork = (work: WorkRecord): UpsertWorkProposal => ({
 });
 
 export interface RemoveWorkProposal {
-	readonly type: "remove_work";
-	readonly workKey: string;
+	type: "remove_work";
+	workKey: string;
 }
 export const removeWork = (key: string): RemoveWorkProposal => ({
 	type: "remove_work",
@@ -193,21 +159,21 @@ export type ReconcileProposal =
 	| RemoveWorkProposal;
 
 export interface WorkItem {
-	readonly workKey: string;
-	readonly subject?: string;
-	readonly templateContext?: unknown;
-	readonly display?: WorkDisplay;
-	readonly operatorActions?: readonly OperatorAction[];
+	workKey: string;
+	subject?: string;
+	templateContext?: unknown;
+	display?: WorkDisplay;
+	operatorActions?: OperatorAction[];
 }
 export interface WorkRun {
-	readonly runId: string;
-	readonly sourceId: string;
-	readonly workKey: string;
-	readonly subject?: string;
-	readonly display?: WorkDisplay;
+	runId: string;
+	sourceId: string;
+	workKey: string;
+	subject?: string;
+	display?: WorkDisplay;
 }
 export interface WorkResult {
-	readonly output?: unknown;
+	output?: unknown;
 }
 export type CompletionStatus =
 	| "succeeded"
@@ -215,28 +181,28 @@ export type CompletionStatus =
 	| "interrupted"
 	| "timed_out";
 export interface Completion {
-	readonly runId: string;
-	readonly sourceId: string;
-	readonly workKey: string;
-	readonly status: CompletionStatus;
-	readonly subject?: string;
-	readonly output?: unknown;
-	readonly error?: string;
+	runId: string;
+	sourceId: string;
+	workKey: string;
+	status: CompletionStatus;
+	subject?: string;
+	output?: unknown;
+	error?: string;
 }
 export interface Diagnostic {
-	readonly level: "info" | "warning" | "error";
-	readonly phase: HookPhase | "act" | "policy";
-	readonly message: string;
-	readonly sourceId?: string;
-	readonly runId?: string;
-	readonly workKey?: string;
+	level: "info" | "warning" | "error";
+	phase: HookPhase | "act" | "policy";
+	message: string;
+	sourceId?: string;
+	runId?: string;
+	workKey?: string;
 }
 export interface ScheduledWake {
-	readonly dueAtMs: number;
-	readonly delayMs: number;
-	readonly reason?: string;
-	readonly workKey?: string;
-	readonly attempt?: number;
+	dueAtMs: number;
+	delayMs: number;
+	reason?: string;
+	workKey?: string;
+	attempt?: number;
 }
 export type WorkSkipReason =
 	| "already_running"
@@ -245,47 +211,47 @@ export type WorkSkipReason =
 	| "capacity_exhausted"
 	| "source_concurrency";
 export interface SkippedWork {
-	readonly workKey: string;
-	readonly sourceId: string;
-	readonly reason: WorkSkipReason;
-	readonly detail?: string;
+	workKey: string;
+	sourceId: string;
+	reason: WorkSkipReason;
+	detail?: string;
 }
 export interface RuntimeSnapshot {
-	readonly tickId: number;
-	readonly facts: ReadonlyMap<string, unknown>;
-	readonly observations: readonly Observation[];
-	readonly completions: readonly Completion[];
-	readonly diagnostics: readonly Diagnostic[];
-	readonly work: ReadonlyMap<string, WorkRecord>;
-	readonly running: ReadonlyMap<string, WorkRun>;
-	readonly scheduledWakes?: readonly ScheduledWake[];
+	tickId: number;
+	facts: Map<string, unknown>;
+	observations: Observation[];
+	completions: Completion[];
+	diagnostics: Diagnostic[];
+	work: Map<string, WorkRecord>;
+	running: Map<string, WorkRun>;
+	scheduledWakes?: ScheduledWake[];
 }
 export interface TickResult {
-	readonly tickId: number;
-	readonly observations: readonly Observation[];
-	readonly proposals: readonly ReconcileProposal[];
-	readonly selected: readonly WorkItem[];
-	readonly started: readonly WorkRun[];
-	readonly skipped: readonly SkippedWork[];
-	readonly completions: readonly Completion[];
-	readonly diagnostics: readonly Diagnostic[];
-	readonly snapshot: RuntimeSnapshot;
+	tickId: number;
+	observations: Observation[];
+	proposals: ReconcileProposal[];
+	selected: WorkItem[];
+	started: WorkRun[];
+	skipped: SkippedWork[];
+	completions: Completion[];
+	diagnostics: Diagnostic[];
+	snapshot: RuntimeSnapshot;
 }
 export type PlotAgentEvent =
-	| { readonly type: "tick_started"; readonly tickId: number }
-	| { readonly type: "tick_completed"; readonly result: TickResult }
-	| { readonly type: "work_observed"; readonly work: WorkRecord }
-	| { readonly type: "work_removed"; readonly workKey: string }
+	| { type: "tick_started"; tickId: number }
+	| { type: "tick_completed"; result: TickResult }
+	| { type: "work_observed"; work: WorkRecord }
+	| { type: "work_removed"; workKey: string }
 	| {
-			readonly type: "wake_scheduled";
-			readonly delayMs: number;
-			readonly reason?: string;
-			readonly workKey?: string;
-			readonly attempt?: number;
+			type: "wake_scheduled";
+			delayMs: number;
+			reason?: string;
+			workKey?: string;
+			attempt?: number;
 	  }
-	| { readonly type: "attempt_started"; readonly run: WorkRun }
-	| { readonly type: "attempt_completed"; readonly completion: Completion };
+	| { type: "attempt_started"; run: WorkRun }
+	| { type: "attempt_completed"; completion: Completion };
 export type PlotAgentMessage =
-	| { readonly type: "tick" }
-	| { readonly type: "observation"; readonly observation: Observation }
-	| { readonly type: "shutdown" };
+	| { type: "tick" }
+	| { type: "observation"; observation: Observation }
+	| { type: "shutdown" };
