+++ packages/agent/src/work-runner.ts
@@ -7,20 +7,16 @@ import type {
 } from "./model.js";
 
 export interface WorkRunnerContext {
-	readonly sourceId: string;
-	readonly tickId: number;
-	readonly run: WorkRun;
-	readonly work: WorkItem;
-	readonly snapshot: RuntimeSnapshot;
-	readonly signal: AbortSignal;
-	readonly emitObservation: (
-		observation: Observation,
-	) => boolean | Promise<boolean>;
-	readonly shouldContinue?: (turnNumber: number) => boolean | Promise<boolean>;
+	sourceId: string;
+	tickId: number;
+	run: WorkRun;
+	work: WorkItem;
+	snapshot: RuntimeSnapshot;
+	signal: AbortSignal;
+	emitObservation: (observation: Observation) => boolean | Promise<boolean>;
+	shouldContinue?: (turnNumber: number) => boolean | Promise<boolean>;
 }
 
 export interface WorkRunner {
-	readonly run: (
-		context: WorkRunnerContext,
-	) => WorkResult | Promise<WorkResult>;
+	run: (context: WorkRunnerContext) => WorkResult | Promise<WorkResult>;
 }
